+++ src/Avm.Authoring/Public/Invoke-AvmTestUnit.ps1
@@ -58,11 +58,15 @@ function Invoke-AvmTestUnit {
 
         [switch] $AllowPathFallback,
 
-        [switch] $NoInit
+        [switch] $NoInit,
+
+        [switch] $SkipModuleVersionCheck
     )
 
     Set-StrictMode -Version 3.0
     $ErrorActionPreference = 'Stop'
+
+    Test-AvmModuleVersion -SkipModuleVersionCheck:$SkipModuleVersionCheck
 
     $context = Get-AvmModuleContext -Path $Path -Ecosystem $Ecosystem